move range summary functions to helper #38
