+++ src/equitycalculator/helpers/general.js
@@ -1,4 +1,10 @@
-import { getSuitFromUnicode, unicodeSuits } from '../'
+import React from 'react'
+import {
+    getSuitFromUnicode,
+    getSuitFromHTML,
+    ranks,
+    unicodeSuits
+} from '../'
 
 export const getCardStringFromObj = (card) => {
     return card.rank + getSuitFromUnicode(card.suit)
@@ -78,4 +84,87 @@ export const formatPercentage = (number, decimalPlaces) => {
         return '<1%'
     }
     return (number * 100).toFixed(decimalPlaces) + '%'
+}
+
+export const getHand = (hand) => {
+    let formattedHand = []
+
+    Object.values(hand).forEach((card, i) => {
+        const suit = getSuitFromHTML(card.suit.charCodeAt())
+        formattedHand.push(
+            <span key={i} className={suit}>
+                {card.rank + card.suit}
+            </span>
+        )
+    })
+
+    return formattedHand
+}
+
+const getPairSummary = (pairs) => {
+    const sortedPairs = pairs.sort((a, b) => ranks.indexOf(a[0]) - ranks.indexOf(b[0]))
+    let pairRanges = [{
+        highest: '',
+        lowest: ''
+    }]
+
+    sortedPairs.forEach((pair, i) => {
+        const currentRange = pairRanges.length - 1
+
+        if (pairRanges[currentRange].highest.length === 0) {
+            pairRanges[currentRange].highest = pair
+        }
+
+        if ((i + 1) === sortedPairs.length) {
+            pairRanges[currentRange].lowest = pair
+        } else if (ranks.indexOf(sortedPairs[i + 1][0]) - ranks.indexOf(sortedPairs[i][0]) !== 1) {
+            pairRanges[currentRange].lowest = pair
+            pairRanges.push({
+                highest: '',
+                lowest: ''
+            })
+        }
+    })
+
+    let pairSummary = ''
+
+    pairRanges.forEach((pairRange, i) => {
+        pairSummary += pairRange.highest
+
+        if (pairRange.highest !== pairRange.lowest) {
+            pairSummary += '-' + pairRange.lowest
+        }
+        if (pairRanges.length === 1 && sortedPairs.indexOf('AA') !== -1 && sortedPairs.length > 1) {
+            pairSummary = pairRange.lowest + '+'
+        }
+        if ((i + 1) < pairRanges.length) {
+            pairSummary += ', '
+        }
+    })
+
+    return pairSummary
+}
+
+export const getRange = (range) => {
+    const pairs = range.filter(hand => hand.length === 2)
+    const nonPairs = range.filter(hand => hand.length === 3)
+    let rangeSummary = getPairSummary(pairs)
+
+    nonPairs.forEach((hand, i) => {
+        if (i === 0 && pairs.length > 0) {
+            rangeSummary += ', '
+        }
+
+        if (rangeSummary.length < 45) {
+            rangeSummary += hand
+
+            if ((i + 1) < nonPairs.length) {
+                rangeSummary += ', '
+            }
+        } else if (rangeSummary.indexOf('...') === -1) {
+            rangeSummary = rangeSummary.slice(0, -2) + '...'
+        }
+    })
+
+    return rangeSummary
 }
